+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/crwrite.cpp
@@ -22,8 +22,8 @@ CommandCrwriteHelp()
     ShowMessages("!crwrite : monitors modification of control registers (CR0 / CR4).\n\n");
 
     ShowMessages("syntax : \t!crwrite [Cr (hex)] [mask Mask (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
-                 "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [stage CallingStage (prepostall)] "
+                 "[buffer PreAllocatedBuffer (hex)] [script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : !crwrite 0\n");