+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/epthook.cpp
@@ -23,9 +23,8 @@ CommandEptHookHelp()
 
     ShowMessages(
         "syntax : \t!epthook [Address (hex)] [pid ProcessId (hex)] [core CoreId (hex)] "
-        "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
-        "[script { Script (string) }] [condition { Condition (hex) }] "
-        "[code { Code (hex) }] \n");
+        "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] [script { Script (string) }] "
+        "[condition { Condition (hex) }] [code { Code (hex) }] \n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : !epthook nt!ExAllocatePoolWithTag\n");