+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/ioout.cpp
@@ -23,8 +23,8 @@ CommandIooutHelp()
                  "instructions.\n\n");
 
     ShowMessages("syntax : \t!ioout [Port (hex)] [pid ProcessId (hex)] "
-                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
-                 "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
+                 "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [stage CallingStage (prepostall)] "
+                 "[buffer PreAllocatedBuffer (hex)] [script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\n");
     ShowMessages("\t\te.g : !ioout\n");