+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/tsc.cpp
@@ -22,8 +22,8 @@ CommandTscHelp()
     ShowMessages("!tsc : monitors execution of rdtsc/rdtscp instructions.\n\n");
 
     ShowMessages("syntax : \t!tsc [pid ProcessId (hex)] [core CoreId (hex)] "
-                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
-                 "[script { Script (string) }] [condition { Condition (hex) }] "
+                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [stage CallingStage (prepostall)] "
+                 "[buffer PreAllocatedBuffer (hex)] [script { Script (string) }] [condition { Condition (hex) }] "
                  "[code { Code (hex) }]\n");
 
     ShowMessages("\n");