+++ hyperdbg/hprdbgkd/code/debugger/core/Debugger.c
@@ -1057,6 +1057,11 @@ DebuggerTriggerEvents(VMM_EVENT_TYPE_ENUM                   EventType,
         // perform the actions
         //
         DebuggerPerformActions(DbgState, CurrentEvent, Context);
+
+        //
+        // Set the current calling stage
+        //
+        DbgState->CurrentCallingStage = VMM_CALLBACK_CALLING_STAGE_INVALID_EVENT_EMULATION;
     }
 
     //
@@ -1149,6 +1154,18 @@ DebuggerPerformRunScript(PROCESSOR_DEBUGGING_STATE * DbgState,
     ACTION_BUFFER                ActionBuffer  = {0};
     SYMBOL                       ErrorSymbol   = {0};
     SCRIPT_ENGINE_VARIABLES_LIST VariablesList = {0};
+
+    //
+    // Fill the action buffer's calling stage
+    //
+    if (DbgState->CurrentCallingStage == VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION)
+    {
+        ActionBuffer.CallingStage = 1;
+    }
+    else
+    {
+        ActionBuffer.CallingStage = 0;
+    }
 
     if (Action != NULL)
     {
@@ -1244,7 +1261,7 @@ DebuggerPerformRunTheCustomCode(PROCESSOR_DEBUGGING_STATE * DbgState, UINT64 Tag
 
     //
     // -----------------------------------------------------------------------------------------------------
-    // Test (Should be removed)
+    // Test
     //
     // LogInfo("%X       Called from : %llx", Tag, Context);
     //
@@ -2070,8 +2087,8 @@ DebuggerParseEventFromUsermode(PDEBUGGER_GENERAL_EVENT_DETAIL EventDetails, UINT
     // it is because using the short-circuiting mechanism with
     // post-events doesn't make sense; it's not supported!
     //
-    if ((EventDetails->EventMode == VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION ||
-         EventDetails->EventMode == VMM_CALLBACK_CALLING_STAGE_ALL_EVENT_EMULATION) &&
+    if ((EventDetails->EventStage == VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION ||
+         EventDetails->EventStage == VMM_CALLBACK_CALLING_STAGE_ALL_EVENT_EMULATION) &&
         EventDetails->EnableShortCircuiting == TRUE)
     {
         ResultsToReturnUsermode->IsSuccessful = FALSE;
@@ -2916,13 +2933,13 @@ DebuggerParseEventFromUsermode(PDEBUGGER_GENERAL_EVENT_DETAIL EventDetails, UINT
     Event->EnableShortCircuiting = EventDetails->EnableShortCircuiting;
 
     //
-    // Set the event mode (pre- post- event)
+    // Set the event stage (pre- post- event)
     //
-    if (EventDetails->EventMode == VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION)
+    if (EventDetails->EventStage == VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION)
     {
         Event->EventMode = VMM_CALLBACK_CALLING_STAGE_POST_EVENT_EMULATION;
     }
-    else if (EventDetails->EventMode == VMM_CALLBACK_CALLING_STAGE_ALL_EVENT_EMULATION)
+    else if (EventDetails->EventStage == VMM_CALLBACK_CALLING_STAGE_ALL_EVENT_EMULATION)
     {
         Event->EventMode = VMM_CALLBACK_CALLING_STAGE_ALL_EVENT_EMULATION;
     }